pyproject.toml
[tool.poetry]
name = "vendor-extension"
version = "1.0.0"
description = "Vendor Table Manager"
authors = ["Developer <dev@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8"
connect-eaas-core = "*"
starlette = "*"
jinja2 = "*"

[tool.poetry.plugins."connect.eaas.ext"]
"webapp" = "extension:VendorTableExtension"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
